.history/backend/app/requirements_20241228233553.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
motor
pydantic
python-dotenv
meta-ai-api
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0.post1
pymongo==4.6.0
meta-ai-api==1.0.0
python-dateutil==2.8.2
bson==0.5.10
aiohttp==3.9.1
typing-extensions==4.8.0
starlette==0.27.0